修改向量类型，匹配向量数据库

+++ src/main/java/com/unfbx/chatgpt/entity/embeddings/Item.java
@@ -9,6 +9,6 @@ import java.util.List;
 @Data
 public class Item implements Serializable {
     private String object;
-    private List<BigDecimal> embedding;
+    private List<Float> embedding;
     private Integer index;
 }
